refactor(pulumi): rename `config.ts` to `action.ts`
Now the file contains only action-based configs.
